feat: added SchedulingResultSchema and event timeline compression

+++ packages/rr-srtf/src/rr_srtf/schemas/scheduling_metrics/scheduling_system_metrics_schema.py
@@ -6,6 +6,6 @@ class SchedulingSystemMetricsSchema(BaseModel):
     total_time: int
     busy_time: float = Field(alias="busy")
     utilization: float = Field(alias="util")
-    throughput_at_window: int = Field(alias="thr_at_window")
+    throughput_at_window: float = Field(alias="thr_at_window")
     throughput_overall: float = Field(alias="thr_overall")
